5.3.5 版本发布

+++ udesksdk/UdeskSDKUI/src/main/java/cn/udesk/aac/UdeskViewMode.java
@@ -140,7 +140,7 @@ public class UdeskViewMode extends ViewModel {
         robotApiData.setCustomerId(customerId);
     }
 
-    public void setSessionId(int sessionId) {
+    public void setSessionId(long sessionId) {
         robotApiData.setSessionId(sessionId);
     }
